+++ src/tracker/platform/psmove_linuxsupport.c
@@ -66,7 +66,7 @@ linux_find_pseye()
                     &capability) == 0, -1);
 
         if (strcmp((const char*)(capability.driver), "ov534") == 0) {
-            psmove_DEBUG("Detected PSEye (ov534): %s\n", filename);
+            PSMOVE_DEBUG("Detected PSEye (ov534): %s", filename);
             result = i;
             break;
         }