+++ src/utils/psmovecli.cpp
@@ -139,7 +139,7 @@ public:
         , done(0)
     {
         char *host_address = psmove_port_get_host_bluetooth_address();
-        printf("Host address: %s\n", host_address ? host_address : "unknown");
+        PSMOVE_INFO("Host address: %s", host_address ? host_address : "unknown");
     }
 
     virtual void connect(Controller *controller) {